requirements.txt
selenium
pytest
pytest-html
faker
pyyaml
python-dotenv